+++ crates/cast/tests/e2e/account/create.rs
@@ -2,9 +2,10 @@ use crate::helpers::constants::{CONTRACTS_DIR, DEVNET_OZ_CLASS_HASH, URL};
 use crate::helpers::fixtures::{default_cli_args, duplicate_directory_with_salt};
 use crate::helpers::runner::runner;
 use camino::Utf8PathBuf;
+use cast::helpers::constants::TEST_KEYSTORE_PASSWORD_ENV_VAR;
 use indoc::indoc;
 use snapbox::cmd::{cargo_bin, Command};
-use std::fs;
+use std::{env, fs};
 
 #[tokio::test]
 pub async fn test_happy_case() {
@@ -233,4 +234,132 @@ pub async fn test_account_already_exists() {
         command: account create
         error: Account with name user1 already exists in network with chain_id SN_GOERLI
     "#});
+}
+
+#[tokio::test]
+pub async fn test_happy_case_keystore() {
+    let keystore_path = "my_key.json";
+    let account_path = "my_account.json";
+    _ = fs::remove_file(keystore_path);
+    _ = fs::remove_file(account_path);
+    env::set_var(TEST_KEYSTORE_PASSWORD_ENV_VAR, "123");
+
+    let args = vec![
+        "--url",
+        URL,
+        "--keystore",
+        keystore_path,
+        "--account",
+        account_path,
+        "account",
+        "create",
+        "--class-hash",
+        DEVNET_OZ_CLASS_HASH,
+    ];
+
+    let snapbox = runner(&args);
+
+    snapbox.assert().stdout_matches(indoc! {r#"
+        TEST_KEYSTORE_PASSWORD environment variable for testing found and will be used
+        Account successfully created[..]
+        command: account create
+        add_profile: --add-profile flag was not set. No profile added to Scarb.toml
+        address: 0x[..]
+        max_fee: 0x[..]
+    "#});
+
+    let contents = fs::read_to_string(account_path).expect("Unable to read created file");
+    assert!(contents.contains("\"deployment\": {"));
+    assert!(contents.contains("\"variant\": {"));
+    assert!(contents.contains("\"version\": 1"));
+
+    _ = fs::remove_file(keystore_path);
+    _ = fs::remove_file(account_path);
+}
+
+#[tokio::test]
+pub async fn test_happy_case_keystore_add_profile() {
+    let current_dir = Utf8PathBuf::from(duplicate_directory_with_salt(
+        CONTRACTS_DIR.to_string() + "/v1/map",
+        "put",
+        "5",
+    ));
+    let keystore_path = "my_key.json";
+    let account_path = "my_account.json";
+    let accounts_file = "accounts.json";
+    env::set_var(TEST_KEYSTORE_PASSWORD_ENV_VAR, "123");
+
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_file,
+        "--keystore",
+        keystore_path,
+        "--account",
+        account_path,
+        "account",
+        "create",
+        "--class-hash",
+        DEVNET_OZ_CLASS_HASH,
+        "--add-profile",
+    ];
+
+    let snapbox = Command::new(cargo_bin!("sncast"))
+        .current_dir(&current_dir)
+        .args(args);
+    let bdg = snapbox.assert().success();
+    let out = bdg.get_output();
+
+    let stdout_str =
+        std::str::from_utf8(&out.stdout).expect("failed to convert command output to string");
+    assert!(stdout_str.contains("add_profile: Profile successfully added to Scarb.toml"));
+
+    let contents =
+        fs::read_to_string(current_dir.join("Scarb.toml")).expect("Unable to read Scarb.toml");
+    assert!(contents.contains("[tool.sncast.my_account]"));
+    assert!(contents.contains("account = \"my_account.json\""));
+
+    let contents =
+        fs::read_to_string(current_dir.join(account_path)).expect("Unable to read created file");
+    assert!(contents.contains("\"deployment\": {"));
+    assert!(contents.contains("\"variant\": {"));
+    assert!(contents.contains("\"version\": 1"));
+
+    let contents =
+        fs::read_to_string(current_dir.join("Scarb.toml")).expect("Unable to read Scarb.toml");
+    assert!(contents.contains(r#"[tool.sncast.my_account]"#));
+    assert!(contents.contains(r#"account = "my_account.json""#));
+    assert!(contents.contains(r#"accounts-file = "accounts.json""#));
+    assert!(contents.contains(r#"keystore = "my_key.json""#));
+    assert!(contents.contains(r#"url = "http://127.0.0.1:5055/rpc""#));
+
+    fs::remove_dir_all(current_dir).unwrap();
+}
+
+#[tokio::test]
+pub async fn test_keystore_without_account() {
+    let keystore_path = "my_key.json";
+    _ = fs::remove_file(keystore_path);
+    env::set_var(TEST_KEYSTORE_PASSWORD_ENV_VAR, "123");
+
+    let args = vec![
+        "--url",
+        URL,
+        "--keystore",
+        keystore_path,
+        "account",
+        "create",
+        "--class-hash",
+        DEVNET_OZ_CLASS_HASH,
+    ];
+
+    let snapbox = runner(&args);
+
+    snapbox.assert().stderr_matches(indoc! {r#"
+        command: account create
+        error: --account must be passed when using --keystore
+    "#});
+
+    _ = fs::remove_file(keystore_path);
 }